リファクタリング

+++ MyGameEngine/Result.h
@@ -6,8 +6,8 @@
 class Result : public GameObject
 {
 	int hModel_;						//モデル番号を格納する変数
-	const int BACK_GROUND_VERTICAL_;	//背景のY軸の高さ
-	const int BACK_GROUND_BESIDE_;		//背景のX軸の高さ
+	const float BACK_GROUND_VERTICAL_;	//背景のY軸の高さ
+	const float BACK_GROUND_BESIDE_;	//背景のX軸の高さ
 	const float BACK_GROUND_DEPTH_;		//背景の奥行
 	const string FILE_PAS_;				//リザルト画面のモデルのファイルパス
 public:
